requirements.txt
bidict==0.21.2
click==7.1.2pip
Flask==1.1.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
SQLAlchemy==1.4.0b1
Werkzeug==1.0.1
WTForms==3.0.0a1
